refactor: split drag drop state logic
